requirements.txt
django==4.2.5
crispy-bootstrap5==0.7
django-filter==23.3
djangorestframework==3.14.0
